# Always log if there's an exception occurred.