Hi,

I'm trying to run sanic in one of my server. My problem is that when there's an exception, I cannot see what happened because there's no log to see. The solution to this is to enable debug mode, the problem with that is the user can see the exception logs as well.

I'm not sure this is the best approach but I propose we always log the exception even though we are not running in debug mode.