@awiddersheim 2017-01-17
I agree with this as well. Thought it was a bit weird when I found out sanic was swallowing all of the exceptions unless in debug.