@odarbelaeze 2017-01-17
Isn't it good to shallow those exceptions for security reasons, maybe you could trow these into a log if not in debug mode and then check the log directly on the server?