@r0fls 2017-01-17
I see no noticeable degradation from these changes:

*After with no exception*
```
$ wrk -t50 -c100 -d30 http://localhost:8000
Running 30s test @ http://localhost:8000
  50 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.45ms  510.97us  15.91ms   82.72%
    Req/Sec     1.39k   272.20     9.94k    73.65%
  2080812 requests in 30.10s, 255.99MB read
Requests/sec:  69129.72
Transfer/sec:      8.50MB
```

*Before with no exception*
```
$ wrk -t50 -c100 -d30 http://localhost:8000
Running 30s test @ http://localhost:8000
  50 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.48ms  519.05us  28.79ms   84.82%
    Req/Sec     1.37k   240.10    16.64k    77.67%
  2049662 requests in 30.10s, 252.16MB read
Requests/sec:  68094.88
Transfer/sec:      8.38MB
```

*After with exception*
```
$ wrk -t50 -c100 -d30 http://localhost:8000
Running 30s test @ http://localhost:8000
  50 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.65ms  618.78us  26.86ms   77.83%
    Req/Sec     1.22k   316.54    14.46k    74.39%
  1827891 requests in 30.10s, 338.18MB read
  Non-2xx or 3xx responses: 1827891
Requests/sec:  60734.48
Transfer/sec:     11.24MB
```

*Before with exception*
```
$ wrk -t50 -c100 -d30 http://localhost:8000
Running 30s test @ http://localhost:8000
  50 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.64ms  509.15us  11.68ms   74.98%
    Req/Sec     1.23k   218.65     2.02k    69.49%
  1832898 requests in 30.02s, 339.11MB read
  Non-2xx or 3xx responses: 1832898
Requests/sec:  61046.20
Transfer/sec:     11.29MB
```