@r0fls 2017-01-17
@zkanda is it possible to write a unit test to make sure it's doing what you expect, or is that too difficult? I had some troubles when testing with logging, since there are some oddities with the pytest logging treatment.