@r0fls 2017-01-17
I ask because I implemented your change and did not see an exception logged in non debug mode when I used an invalid route:

```python
from sanic.response import text
from sanic import Sanic
import asyncio
import time

app = Sanic()

@app.route('/')
async def hello(request):
    x == 3
    return text("OK")

if __name__ == '__main__':
    app.run(host="0.0.0.0",
            port=8000,
            workers=4)
```