@r0fls 2017-01-17
@zkanda that's ok if that's a different bug, but I would need to see an example where this does something before approving it. Can you provide one?