@GenericError 2017-01-18
This would bring it into line with similar programs - errors are generally at the default level of output