@zkanda 2017-01-18
Seems like something is wrong with the test..

@r0fls how are you running the test? I'm using tox but I would prefer to use py.test directly.

* Here's how I did it, install all requirements-dev.txt
* Run `pytest`

I get a lot of failed test. 90 in total

e.g.

```
tests/test_views.py:194:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

app = <sanic.sanic.Sanic object at 0x7f18d426a828>, method = 'get', uri = '/', gather_request = True, loop = None, debug = False, server_kwargs = {}, request_args = (), request_kwargs = {}
_collect_request = <function sanic_endpoint_test.<locals>._collect_request at 0x7f18d6713950>, _collect_response = <function sanic_endpoint_test.<locals>._collect_response at 0x7f18d42da048>

    def sanic_endpoint_test(app, method='get', uri='/', gather_request=True,
                            loop=None, debug=False, server_kwargs={},
                            *request_args, **request_kwargs):
        results = []
        exceptions = []

        if gather_request:
            def _collect_request(request):
                results.append(request)
            app.request_middleware.appendleft(_collect_request)

        async def _collect_response(sanic, loop):
            try:
                response = await local_request(method, uri, *request_args,
                                               **request_kwargs)
                results.append(response)
            except Exception as e:
                exceptions.append(e)
            app.stop()

        app.run(host=HOST, debug=debug, port=PORT,
                after_start=_collect_response, loop=loop, **server_kwargs)

        if exceptions:
            raise ValueError("Exception during request: {}".format(exceptions))

        if gather_request:
            try:
                request, response = results
                return request, response
            except:
                raise ValueError(
                    "Request and response object expected, got ({})".format(
>                       results))
E               ValueError: Request and response object expected, got ([])

../../../.virtualenvs/sanic/lib/python3.6/site-packages/Sanic-0.2.0-py3.6.egg/sanic/utils.py:51: ValueError
================================================================================== 90 failed, 8 passed, 2 skipped in 6.87 seconds ===================================================================================
```