@r0fls 2017-01-18
@zkanda I'm able to run the tests with `python3 -m pytest tests` from within the directory. Is that the command you're using? Anyway I'm fine merging this as is, because I agree it's the right behavior.