@zkanda 2017-01-19
hmmm the reason why it's failing on test is because the `request` object is `None` when we call `request.url`. I tried debugging why it's not passing that object properly on testing but I'm not so familiar with the code base yet(and currently at work).

With that said, if we just log without the request then obviously the test would pass but there is redundancy in logging if we are in debug mode.