@r0fls 2017-01-19
@zkanda let's start here, if you want to improve the functionality further just open another PR. This is an improvement already since if there is an exception, it should probably be logged.